% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdi_likelihood_surface.R
\name{bdi_lrt_moving_window}
\alias{bdi_lrt_moving_window}
\title{Moving window}
\usage{
bdi_lrt_moving_window(z, delta_t = 1, eta = 1, gamm = 1, dR0 = 0,
  R00 = 0.9, fp_e = 2, fp_s = 1, window = 10)
}
\description{
Moving window
}
